view transform

+++ src/main/scala/org/bbstilson/raytracer/transforms/Translation.scala
@@ -11,8 +11,8 @@ case class Translation(m: Matrix) extends MatrixDouble(m) {
 
 object Translation {
 
-  def apply(x: Int, y: Int, z: Int): Translation = {
-    val vals = Vector(x.toDouble, y.toDouble, z.toDouble, 1d)
+  def apply(x: Double, y: Double, z: Double): Translation = {
+    val vals = Vector(x, y, z, 1)
     val f = (r: Int, c: Int) => if (r == c) 1d else if (c == 3) vals(r) else 0d
     val m = mkMatrix(4, 4, f)
     new Translation(m)
